Edits with the aim of generalising the check_cube_not_float64 function to be a check_cube_datatypes function, which corrects the cube data and coordinates to fit with the expectations in IMPROVER. Unit tests have been added and some modifications of other unit tests were required related to stricter enforcement of datatypes within cubes.

+++ lib/improver/tests/utilities/test_load.py
@@ -182,7 +182,7 @@ class Test_load_cube(IrisTest):
         cube = set_up_probability_cube(
             np.zeros((3, 4, 5), dtype=np.float32),
             np.array([273., 274., 275.], dtype=np.float32))
-        cube = add_coordinate(cube, [0, 1, 2], "realization")
+        cube = add_coordinate(cube, [0, 1, 2], "realization", dtype=np.int32)
         cube = add_coordinate(
             cube, [10, 50, 90], "percentile_over_neighbourhood")
         cube.transpose([4, 3, 2, 1, 0])
